syntax: Extract `Parser` monad into its own module
